# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Test API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/v1/invoices": {
      "get": {
        "tags": [
          "Invoices"
        ],
        "summary": "Get all invoices",
        "description": "Retrieves a list of invoices with optional search and sorting",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Search term for invoice number or user details",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Field to sort by (default: issuedAt)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort order (default: desc)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of invoices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Invoice"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "tags": [
          "Invoices"
        ],
        "summary": "Create a new invoice",
        "description": "Creates a new invoice with associated items and calculates totals",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invoice created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/api/v1/items/{id}": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Get item by ID",
        "description": "Retrieves a specific item with details",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "put": {
        "tags": [
          "Items"
        ],
        "summary": "Update item by ID",
        "description": "Updates an existing item",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "tags": [
          "Items"
        ],
        "summary": "Delete item by ID",
        "description": "Deletes a specific item if it is not referenced in any invoices",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/api/v1/items/generate-number": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Generate a unique item number",
        "description": "Generates a unique item number",
        "responses": {
          "200": {
            "description": "Generated item number",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "itemNumber": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "example": "no-store, max-age=0"
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/api/v1/item-taxes": {
      "get": {
        "tags": [
          "ItemTaxes"
        ],
        "summary": "Get all item taxes",
        "description": "Retrieves a list of item taxes with their labels",
        "responses": {
          "200": {
            "description": "List of item taxes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "taxes": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ItemTaxLabel"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userNumber": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string"
          },
          "vatNumber": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "SuccessMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "Invoice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "invoiceNumber": {
            "type": "string"
          },
          "issuedAt": {
            "type": "string",
            "format": "date-time"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time"
          },
          "totalAmount": {
            "type": "number",
            "format": "float"
          },
          "totalVatAmount": {
            "type": "number",
            "format": "float"
          },
          "totalTtcAmount": {
            "type": "number",
            "format": "float"
          },
          "flag_accounting": {
            "type": "boolean"
          },
          "User": {
            "$ref": "#/components/schemas/User"
          },
          "invoiceDetails": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceDetail"
            }
          }
        }
      },
      "InvoiceDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "lineNumber": {
            "type": "integer"
          },
          "quantity": {
            "type": "integer"
          },
          "discount": {
            "type": "number",
            "format": "float"
          },
          "unitPrice": {
            "type": "number",
            "format": "float"
          },
          "vatBaseAmount": {
            "type": "number",
            "format": "float"
          },
          "vatAmount": {
            "type": "number",
            "format": "float"
          },
          "totalPrice": {
            "type": "number",
            "format": "float"
          },
          "item": {
            "$ref": "#/components/schemas/Item"
          }
        }
      },
      "InvoiceCreate": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceItemCreate"
            }
          },
          "issuedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "userId",
          "items",
          "issuedAt"
        ]
      },
      "InvoiceItemCreate": {
        "type": "object",
        "properties": {
          "itemId": {
            "type": "string"
          },
          "quantity": {
            "type": "integer"
          },
          "discount": {
            "type": "number",
            "format": "float"
          }
        },
        "required": [
          "itemId",
          "quantity"
        ]
      },
...
      "Item": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "itemNumber": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "retailPrice": {
            "type": "number",
            "format": "float"
          },
          "purchasePrice": {
            "type": "number",
            "format": "float"
          },
          "stockQuantity": {
            "type": "integer"
          },
          "minQuantity": {
            "type": "integer"
          },
          "vat": {
            "$ref": "#/components/schemas/Vat"
          },
          "unit": {
            "$ref": "#/components/schemas/Unit"
          },
          "itemClass": {
            "$ref": "#/components/schemas/ItemClass"
          },
          "itemTaxes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemTax"
            }
          }
        }
      },
...
      "ItemUpdate": {
        "type": "object",
        "properties": {
          "itemNumber": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "retailPrice": {
            "type": "number",
            "format": "float"
          },
          "purchasePrice": {
            "type": "number",
            "format": "float"
          },
          "vatId": {
            "type": "integer"
          },
          "unitId": {
            "type": "integer"
          },
          "classId": {
            "type": "integer"
          },
          "itemTaxes": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          }
        },
        "required": [
          "label",
          "retailPrice",
          "purchasePrice",
          "vatId",
          "unitId",
          "classId"
        ]
      },
      "Vat": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "vatPercent": {
            "type": "number",
            "format": "float"
          },
          "vatType": {
            "type": "string"
          },
          "country": {
            "$ref": "#/components/schemas/Country"
          }
        }
      },
      "Country": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "countryCode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "ItemTax": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "itemId": {
            "type": "integer"
          },
          "taxId": {
            "type": "integer"
          },
          "price": {
            "type": "number",
            "format": "float"
          },
          "utax": {
            "$ref": "#/components/schemas/Utax"
          }
        }
      },
      "Utax": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "rate": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "ItemClass": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "label": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          }
        }
      },
      "ItemTaxLabel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "label": {
            "type": "string"
          }
        }
      },
      "Unit": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "404": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}